+++ src/relatedProperties/propertyTypes/PropertyTypeList.php
@@ -372,7 +372,7 @@ class PropertyTypeList extends PropertyType
     /**
      * @return string
      */
-    public function getStringValue()
+    public function getAsText()
     {
         $value = $this->property->relatedPropertiesModel->getAttribute($this->property->code);