+++ src/relatedProperties/propertyTypes/PropertyTypeTree.php
@@ -132,7 +132,7 @@ class PropertyTypeTree extends PropertyType
     /**
      * @return string
      */
-    public function getStringValue()
+    public function getAsText()
     {
         $value = $this->property->relatedPropertiesModel->getAttribute($this->property->code);